refactor(strategy): 전략 패턴 패키지 분리 vo -> strategy
