+++ src/main/java/racingcar/domain/strategy/MovableStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.domain.vo;
+package racingcar.domain.strategy;
 
 public interface MovableStrategy {
     boolean isMove();
